requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
python-dotenv
python-jose
passlib[bcrypt]
pydantic
